api-django/requirements.txt
Django==2.0.3
django-filter==1.1.0
djangorestframework==3.7.7
mysqlclient==1.3.12
drf-yasg[validation]==1.5.1
django-debug-toolbar==1.9.1